refactor: controllers, array method
