+++ task5/src/controllers/index.ts
@@ -1,8 +1,3 @@
-export { default as createUserController } from "./createUserController";
-export { default as getUsersController } from "./getUsersController";
-export { default as getUserController } from "./getUserController";
-export { default as getUserHobbiesController } from "./getUserHobbiesController";
-export { default as deleteUserController } from "./deleteUserController";
-export { default as updateUserHobbyController } from "./updateUserHobbyController";
-export { default as updateUserController } from "./updateUserController";
 export { default as mainController } from "./mainController";
+export { default as userController } from "./userController";
+export { default as userHobbyController } from "./userHobbyController";
